endpoint/requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
anyio==3.7.1
appnope==0.1.3
arrow==1.3.0
astrapy==0.5.1
asttokens==2.4.0
async-timeout==4.0.3
attrs==23.1.0
backcall==0.2.0
Brotli==1.1.0
cassandra-driver==3.28.0
cassio==0.1.1
certifi==2023.7.22
chardet==5.2.0
charset-normalizer==3.2.0
click==8.1.7
comm==0.1.4
coverage==7.3.2
dataclasses-json==0.5.14
datasets==2.14.5
debugpy==1.8.0
decorator==5.1.1
deepgram-sdk==2.11.0
dill==0.3.7
distro==1.8.0
docopt==0.6.2
exceptiongroup==1.1.3
executing==1.2.0
Faker==19.11.1
fastapi==0.103.1
filelock==3.12.4
frozenlist==1.4.0
fsspec==2023.6.0
geomet==0.2.1.post1
greenlet==2.0.2
h11==0.14.0
httpcore==1.0.2
httptools==0.6.0
httpx==0.25.2
huggingface-hub==0.17.2
idna==3.4
importlib-metadata==6.8.0
iniconfig==2.0.0
ipykernel==6.25.2
ipython==8.15.0
jedi==0.19.0
jupyter_client==8.6.0
jupyter_core==5.5.0
langchain==0.0.295
langsmith==0.0.38
marshmallow==3.20.1
matplotlib-inline==0.1.6
multidict==6.0.4
multiprocess==0.70.15
mutagen==1.47.0
mypy-extensions==1.0.0
nest-asyncio==1.5.8
numexpr==2.8.6
numpy==1.26.0
openai==1.3.6
packaging==23.1
pandas==2.1.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==3.10.0
pluggy==1.3.0
prompt-toolkit==3.0.39
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==13.0.0
pycryptodomex==3.19.0
pydantic==2.3.0
pydantic_core==2.6.3
Pygments==2.16.1
pysrt==1.1.2
pytest==7.4.3
pytest-cov==4.1.0
pytest-testdox==3.1.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.1
regex==2023.8.8
requests==2.31.0
requests-toolbelt==1.0.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.21
stack-data==0.6.2
starlette==0.27.0
tenacity==8.2.3
tiktoken==0.5.1
tomli==2.0.1
tornado==6.3.3
tqdm==4.66.1
traitlets==5.10.0
types-python-dateutil==2.8.19.14
typing-inspect==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.4
uvicorn==0.23.2
uvloop==0.17.0
watchfiles==0.20.0
wcwidth==0.2.6
websockets==11.0.3
webvtt-py==0.4.6
xxhash==3.3.0
yarl==1.9.2
youtube-dl==2021.12.17
yt-dlp==2023.11.16
zipp==3.17.0